+++ docs/md/Version-history.md
@@ -14,6 +14,9 @@ Available in [pre-release](https://www.sumatrapdfreader.org/prerelease) builds.
 - add `CmdDocumentExtractText` (`Extract Text From Document` in `Ctrl + k` [command palette](Command-Palette.md)) to extract text from document pages to a .txt file
 - add `ToolbarText` parameter for `ExternalViewers` advanced setting to show external viewer as a toolbar button
 - move `Scrollbars` advanced setting from `FixedPageUI` to top-level
+- add `EBookUI.BackgroundColor` advanced setting to override background color for ebook documents (epub, mobi etc.)
+- add `ComicBookUI.BackgroundColor` advanced setting to override the default black background for comic book files
+- add `ImageUI.BackgroundColor` advanced setting to override the default black background for image files
 - `Ctrl + click` on a PDF link opens it in a new tab (instead of navigating in the current tab)
 - you can now drag&drop selected text to another application, like a text editor
 - added `List Printers` (`CmdListPrinters`) command to `Ctrl + k` Command Palette to list available printers